Move things from BinaryOp
