+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/design/SimplifyConditionalRule.java
@@ -8,9 +8,9 @@ import static net.sourceforge.pmd.lang.java.ast.BinaryOp.CONDITIONAL_AND;
 import static net.sourceforge.pmd.lang.java.ast.BinaryOp.CONDITIONAL_OR;
 import static net.sourceforge.pmd.lang.java.ast.BinaryOp.INSTANCEOF;
 import static net.sourceforge.pmd.lang.java.ast.BinaryOp.NE;
-import static net.sourceforge.pmd.lang.java.ast.BinaryOp.isInfixExprWithOperator;
 import static net.sourceforge.pmd.lang.java.ast.internal.JavaAstUtils.getOtherOperandIfInInfixExpr;
 import static net.sourceforge.pmd.lang.java.ast.internal.JavaAstUtils.isBooleanNegation;
+import static net.sourceforge.pmd.lang.java.ast.internal.JavaAstUtils.isInfixExprWithOperator;
 import static net.sourceforge.pmd.lang.java.rule.internal.JavaRuleUtil.isNullCheck;
 
 import net.sourceforge.pmd.lang.java.ast.ASTExpression;
